cult-of-aggeris-readme
Hey, welcome to our humble little project that we hope to grow into something a bit more than what Starbound was or became. We aren't much in skill but we do make up for it in pleasing art and steady growth! So if anything caught your attention within our project, please drop by our discord: https://discord.gg/8jfYtJF it would be great if you could, so we can feel more motivated outside of our numbers on steam!